Remove abstraction of --driver
Previously, a user didn't know what --driver was inputing into soapysdr
without reading the source code. Now, the full driver= string is printed
to the user, as the same string they would use with SoapySDRUtil.

This also prints the driver= to the user when a config key values isn't
found, hopefully leading them to the light off adding a config file
option.